Description
This program can general knowledge questions in a JSON file(you can add or remove data from the file). It will generate 5 random questions from the saved question in the JSON file. In the JSON file, the correct answers are also be inputted. If the answer is correct, it will count point 1 for each corrected answer. For each wrong answer, the point is 0. After ending the quiz, it will show a total of points out of 5 questions.